pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-evaluation-toolkit"
version = "0.1.0"
authors = [{name = "Your Name"}]
description = "A toolkit for evaluating LLM outputs"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "sentence-transformers>=2.2.0",
    "transformers>=4.30.0",
    "python-Levenshtein>=0.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black", "flake8"]